Support for different content widths and heights.
User can specify the List<double> contentCellWidths and List<double> contentCellHeights in CellDimensions to create a table with different content cell widths and heights. The contentCellWidth and contentCellHeight parameters, as well as the base values still work as before, so the update is backwards compatible. There are checks so that the user has to specify either contentCellWidths or contentCellWidth, and either contentCellHeights or contentCellHeight, if setting the Cell Dimensions. Also, if setting contentCellWidths, the length of the list should match columnsLength and the length of contentCellHeights should match rowsLength. If both contentCellWidths and contentCellWidth are specified, the list takes priority. Same for contentCellHeights.